requirements.txt
asgiref==3.3.1
autopep8==1.5.5
boto==2.49.0
boto3==1.17.7
botocore==1.20.7
certifi==2020.12.5
chardet==3.0.4
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
Pillow==8.1.0
pycodestyle==2.6.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.24.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
